+++ test/quack_db/integration/type_matrix_test.exs
@@ -103,7 +103,7 @@ defmodule QuackDB.Integration.TypeMatrixTest do
              "550e8400-e29b-41d4-a716-446655440000",
              "hello",
              "duck",
-             <<3, 245>>
+             "10101"
            ]
   end